requirements.txt
requests>=2.25.0
pyyaml>=5.4.1
click>=8.0.0
python-dateutil>=2.8.1
jsonschema>=3.2.0
